# Dashboard Performance: Memoize widgets