--- a/client/app/components/dashboards/DashboardGrid.jsx
+++ b/client/app/components/dashboards/DashboardGrid.jsx
@@ -33,6 +33,53 @@ const WidgetType = PropTypes.shape({
 const SINGLE = "single-column";
 const MULTI = "multi-column";
 
+const DashboardWidget = React.memo(
+  function DashboardWidget({
+    widget,
+    dashboard,
+    onLoadWidget,
+    onRefreshWidget,
+    onRemoveWidget,
+    onParameterMappingsChange,
+    canEdit,
+    isPublic,
+    isLoading,
+    filters,
+  }) {
+    const { type } = widget;
+    const onLoad = () => onLoadWidget(widget);
+    const onRefresh = () => onRefreshWidget(widget);
+    const onDelete = () => onRemoveWidget(widget.id);
+
+    if (type === WidgetTypeEnum.VISUALIZATION) {
+      return (
+        <VisualizationWidget
+          widget={widget}
+          dashboard={dashboard}
+          filters={filters}
+          canEdit={canEdit}
+          isPublic={isPublic}
+          isLoading={isLoading}
+          onLoad={onLoad}
+          onRefresh={onRefresh}
+          onDelete={onDelete}
+          onParameterMappingsChange={onParameterMappingsChange}
+        />
+      );
+    }
+    if (type === WidgetTypeEnum.TEXTBOX) {
+      return <TextboxWidget widget={widget} canEdit={canEdit} isPublic={isPublic} onDelete={onDelete} />;
+    }
+    return <RestrictedWidget widget={widget} />;
+  },
+  (prevProps, nextProps) =>
+    prevProps.widget === nextProps.widget &&
+    prevProps.canEdit === nextProps.canEdit &&
+    prevProps.isPublic === nextProps.isPublic &&
+    prevProps.isLoading === nextProps.isLoading &&
+    prevProps.filters === nextProps.filters
+);
+
 class DashboardGrid extends React.Component {
   static propTypes = {
     isEditing: PropTypes.bool.isRequired,
@@ -203,38 +250,29 @@ class DashboardGrid extends React.Component {
           onLayoutChange={this.onLayoutChange}
           onBreakpointChange={this.onBreakpointChange}
           breakpoints={{ [MULTI]: cfg.mobileBreakPoint, [SINGLE]: 0 }}>
-          {widgets.map(widget => {
-            const widgetProps = {
-              widget,
-              filters,
-              isPublic,
-              canEdit: dashboard.canEdit(),
-              onDelete: () => onRemoveWidget(widget.id),
-            };
-            const { type } = widget;
-            return (
-              <div
-                key={widget.id}
-                data-grid={DashboardGrid.normalizeFrom(widget)}
-                data-widgetid={widget.id}
-                data-test={`WidgetId${widget.id}`}
-                className={cx("dashboard-widget-wrapper", {
-                  "widget-auto-height-enabled": this.autoHeightCtrl.exists(widget.id),
-                })}>
-                {type === WidgetTypeEnum.VISUALIZATION && (
-                  <VisualizationWidget
-                    {...widgetProps}
-                    dashboard={dashboard}
-                    onLoad={() => onLoadWidget(widget)}
-                    onRefresh={() => onRefreshWidget(widget)}
-                    onParameterMappingsChange={onParameterMappingsChange}
-                  />
-                )}
-                {type === WidgetTypeEnum.TEXTBOX && <TextboxWidget {...widgetProps} />}
-                {type === WidgetTypeEnum.RESTRICTED && <RestrictedWidget widget={widget} />}
-              </div>
-            );
-          })}
+          {widgets.map(widget => (
+            <div
+              key={widget.id}
+              data-grid={DashboardGrid.normalizeFrom(widget)}
+              data-widgetid={widget.id}
+              data-test={`WidgetId${widget.id}`}
+              className={cx("dashboard-widget-wrapper", {
+                "widget-auto-height-enabled": this.autoHeightCtrl.exists(widget.id),
+              })}>
+              <DashboardWidget
+                dashboard={dashboard}
+                widget={widget}
+                filters={filters}
+                isPublic={isPublic}
+                isLoading={widget.loading}
+                canEdit={dashboard.canEdit()}
+                onLoadWidget={onLoadWidget}
+                onRefreshWidget={onRefreshWidget}
+                onRemoveWidget={onRemoveWidget}
+                onParameterMappingsChange={onParameterMappingsChange}
+              />
+            </div>
+          ))}
         </ResponsiveGridLayout>
       </div>
     );