## What type of PR is this? (check all applicable)
<!-- Please leave only what's applicable -->

- [x] Other

## Description
The idea of this PR is to handle this:
| Before | After |
| --------- | ------ |
| ![Screenshot from 2020-03-13 14-41-20](https://user-images.githubusercontent.com/3356951/76645919-ba83df80-6538-11ea-97c7-c748793e01db.png)  |  ![Screenshot from 2020-03-13 14-22-40](https://user-images.githubusercontent.com/3356951/76645734-56f9b200-6538-11ea-80d3-fc52040c954b.png) |

Basically, whenever a widget updates it updates the Dashboard and all children components get re-rendered. This PR memoizes widgets and only update them when one of their props change (need to test a bit more to make sure it does not blow anything). While I also wanted to memoize the base child from ReactGridLayout and avoid the whole grid item to re-render, I couldn't find a working way to do that.

For smaller dashboards I didn't see a relevant difference in the scripting time, but for the React Migration dashboard it did shave off extra 2 seconds in the Scripting Time (14s to 12s avg, tested with #4724, #4726 and this PR changes together).

## Related Tickets & Documents
#4714 

## Mobile & Desktop Screenshots/Recordings (if there are UI changes)
--